pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ae-in-workflow"
version = "0.1.1"
description = "In-workflow Nodes for ComfyUI - Heavy interaction and streaming capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ae", email = "ae@ae-maker.com"},
]
maintainers = [
    {name = "ae", email = "ae@ae-maker.com"},
]
keywords = ["comfyui", "nodes", "workflow", "interactive", "pose", "editor"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
requires-python = ">=3.8"
dependencies = [
    "pygame>=2.0.0",
    "torch>=1.9.0",
    "numpy>=1.19.0",
    "Pillow>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/ae-maker/ae-in-workflow"
Repository = "https://github.com/ae-maker/ae-in-workflow"
Issues = "https://github.com/ae-maker/ae-in-workflow/issues"
Changelog = "https://github.com/ae-maker/ae-in-workflow/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["ae_in_workflow"]
package-dir = {"" = "."}

[tool.setuptools.package-data]
"ae_in_workflow" = ["web/*"]

[project.optional-dependencies]
monitor = ["screeninfo>=0.8.0"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
